pyproject.toml
[project]
name = "cucumber-agent"
version = "0.1.0"
description = "🥒 A clean, modular AI agent framework"
requires-python = ">=3.14"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "CucumberTeam" }]
keywords = ["ai", "agent", "cli", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "httpx>=0.28.0",
    "msgspec>=0.18.0",
    "pydantic>=2.10.0",
    "structlog>=25.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.52",
    "trafilatura>=2.0.0",
    "tiktoken>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pyright>=1.1.0",
    "ruff>=0.9.0",
]
sqlite = []

[project.scripts]
cucumber = "cucumber_agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
pythonVersion = "3.14"
include = ["src"]
exclude = ["tests/**"]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pyright>=1.1.409",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]
